requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==0.16
Jinja2==2.7.1
MarkupSafe==0.18
SQLAlchemy==0.7.9
Tempita==0.5.1
Werkzeug==0.9.4
decorator==3.4.0
google-api-python-client==1.2
httplib2==0.8
itsdangerous==0.23
pbr==0.5.23
requests==2.1.0
sqlalchemy-migrate==0.8.2
wsgiref==0.1.2
gunicorn==0.17.2
psycopg2==2.5
